agent311/pyproject.toml
[project]
name = "agent311"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk",
    "fastapi",
    "matplotlib>=3.10.8",
    "uvicorn",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pyjwt>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["agent311"]
